inline `debug?` call inside `dev-setup`
Inlining the function call shaves 65kb of the minified output on a clean
project with only the `+devtools` option.

The size of the non-inlined version matches exactly that of the inlined version
with the `goog.DEBUG` flag excluded. Most likely, the indirection is preventing
Google Closure from eliminating the dead code from the final build.

The `debug?` function could still be useful for other purposes, but a comment
regarding this *feature* would be nice.